+++ packages/design-system/src/components/NumericInputField/NumericInput.js
@@ -9,12 +9,15 @@ const cx = classNames.bind(styles);
 const baseClass = 'numeric-input';
 
 class NumericInput extends React.PureComponent {
+  hasMin = () => this.props.min !== undefined;
+  hasMax = () => this.props.max !== undefined;
+  
   componentDidMount() {
-    if (this.props.max && parseInt(this.props.value, 10) > this.props.max) {
+    if (this.hasMax() && parseInt(this.props.value, 10) > this.props.max) {
       this.callOnChange(this.props.max);
     }
 
-    if (this.props.min && parseInt(this.props.value, 10) < this.props.min) {
+    if (this.hasMin() && parseInt(this.props.value, 10) < this.props.min) {
       this.callOnChange(this.props.min);
     }
   }
@@ -56,9 +59,9 @@ class NumericInput extends React.PureComponent {
   changeValue = val => {
     if (this.props.value !== '' && this.props.value !== '-') {
       this.callOnChange(this.calcValue(parseInt(this.props.value, 10) + val));
-    } else if (this.props.min && val < this.props.min) {
+    } else if (this.hasMin() && val < this.props.min) {
       this.callOnChange(this.props.min);
-    } else if (this.props.max && val > this.props.max) {
+    } else if (this.hasMax() && val > this.props.max) {
       this.callOnChange(this.props.max);
     } else {
       this.callOnChange(val);
@@ -96,15 +99,19 @@ class NumericInput extends React.PureComponent {
   calcValue = val => {
     const { max, min } = this.props;
 
-    if (max && val > max) {
+    if (this.hasMax() && val > max) {
       return max;
     }
 
-    if (min && val < min) {
+    if (this.hasMin() && val < min) {
       return min;
     }
     return val;
   };
+
+  hasReachedTheLimit = (value, margin) => (
+    margin !== undefined && parseInt(value, 10) === margin
+  ) 
 
   handleIncrementClick = () => {
     this.changeValue(1);
@@ -146,6 +153,7 @@ class NumericInput extends React.PureComponent {
       className
     );
 
+
     return (
       <div className={mergedClassNames} style={this.getComponentStyles()}>
         <input
@@ -156,15 +164,15 @@ class NumericInput extends React.PureComponent {
           onChange={this.handleChange}
           onFocus={this.handleFocus}
           onBlur={this.handleBlur}
+          min={min}
+          max={max}
         />
         {!noControls && (
           <React.Fragment>
             <button
               tabIndex="-1"
               disabled={
-                disabled ||
-                (this.props.max &&
-                  parseInt(this.props.value, 10) === this.props.max)
+                disabled || this.hasReachedTheLimit(value, max)
               }
               onClick={this.handleIncrementClick}
               aria-label="Increment value"
@@ -174,9 +182,7 @@ class NumericInput extends React.PureComponent {
             <button
               tabIndex="-1"
               disabled={
-                disabled ||
-                (this.props.min &&
-                  parseInt(this.props.value, 10) === this.props.min)
+                disabled || this.hasReachedTheLimit(value, min)
               }
               aria-label="Decrement value"
               className={styles[`${baseClass}__decrement`]}
@@ -184,7 +190,7 @@ class NumericInput extends React.PureComponent {
               type="button"
             />
           </React.Fragment>
-        )}
+          )}
       </div>
     );
   }